+++ Projects/Feature/Scenes/SetlistScene/ViewModel/SetlistViewModel.swift
@@ -11,23 +11,36 @@ import Core
 
 final class SetlistViewModel: ObservableObject {
   @Published var isBookmarked: Bool
-  @Published var isEmptySetlist: Bool
-
-  func isBookmark(_ concertInfo: [ArchivedConcertInfo], _ setlist: Setlist?) {
-    self.isBookmarked = {
-      for i in 0..<concertInfo.count {
-        if concertInfo[i].setlist.setlistId == setlist?.id { return true }
-      }
-      return false
-    }()
-  }
-
+//  @Published var isEmptySetlist: Bool
+  @Published var showModal: Bool
+  
+  let koreanConverter = KoreanConverter.shared
+  let dataService = SetlistDataService.shared
+  let dataManager = SwiftDataManager()
+  
+  //  func isBookmark(_ concertInfo: [ArchivedConcertInfo], _ setlist: Setlist?) {
+  //    self.isBookmarked = {
+  //      for i in 0..<concertInfo.count {
+  //        if concertInfo[i].setlist.setlistId == setlist?.id { return true }
+  //      }
+  //      return false
+  //    }()
+  //  }
+  
   var setlistSongName: [String] = []
   var setlistSongKoreanName: [String] = []
   var setlistArtistName: [String] = []
   
   init() {
     self.isBookmarked = false
-    self.isEmptySetlist = false
+//    self.isEmptySetlist = false
+    self.showModal = false
+  }
+  
+  func isEmptySetlist(_ setlist: Setlist) -> Bool {
+    if setlist.sets?.setsSet?.isEmpty == true {
+      return true
+    }
+    return false
   }
 }
